Move all the executable code into the library
